feat(v1): update for sentry
